Revamped emailing system, removed unused requirements, made user alerting faster
