+++ .gitignore
@@ -1,0 +1,4 @@
+.vscode/
+__pycache__/
+creds.json
+scriptDetails.txt
