+++ requirements.txt
@@ -1,27 +1,14 @@
-asn1crypto==0.24.0
-cachetools==2.1.0
-certifi==2018.4.16
-cffi==1.11.5
-chardet==3.0.4
-cryptography==2.2.2
-google-api-python-client==1.7.3
-google-auth==1.5.0
-google-auth-httplib2==0.0.3
-gspread==0.3.0
-httplib2==0.11.3
-idna==2.7
-mailchimp3==3.0.4
-mailer==0.8.1
-oauth2client==4.1.2
-pyasn1==0.4.3
-pyasn1-modules==0.2.2
-pycparser==2.18
-pygsheets==1.1.4
-pyOpenSSL==18.0.0
-python-firebase==1.2
-requests==2.19.1
-rsa==3.4.2
-sheetsu==1.1.0
-six==1.11.0
-uritemplate==3.0.0
-urllib3==1.23
+# G spread
+aiohttp==3.5.1 
+gspread==3.1.0 
+oauth2client==3.0.0
+requests==2.11.1
+
+# Mailchimp
+mailchimp3==3.0.6 
+
+# Firebase
+Pyrebase==3.0.27
+pycryptodome==3.7.2
+requests-toolbelt==0.7.0
+requests==2.11.1
